+++ tests/testthat/test_RecursiveForecaster.R
@@ -395,3 +395,17 @@ test_that("RecursiveForecaster provides a default fallback", {
   expect_number(rr$aggregate())
   expect_true(nrow(rr$errors) > 0L)
 })
+
+test_that("RecursiveForecaster propagates predict parameters changed after training", {
+  task = tsk("airpassengers")
+  split = partition(task, ratio = 0.9)
+  learner = RecursiveForecaster$new(lrn("regr.debug"), lags = 1:3)
+  learner$train(task, split$train)
+  expect_prediction(learner$predict(task, split$test))
+
+  learner$param_set$values$regr.debug.error_predict = 1
+  expect_error(learner$predict(task, split$test))
+
+  learner$param_set$values$regr.debug.error_predict = 0
+  expect_prediction(learner$predict(task, split$test))
+})